refactor(types): change hp/armor/damage to int32_t, add game limits
- Add DMCP_PLAYER_MAX_HEALTH, DMCP_PLAYER_MAX_ARMOR, DMCP_DAMAGE_MAX constants
- Add dmcp_gamemode_t and dmcp_gamemission_t enums
- Change hp/armor in dmcp_player_t from float to int32_t
- Change hp/max_hp in dmcp_enemy_t from float to int32_t
- Change health/damage in command structs from float to int32_t
- Update validation.h to use constants instead of magic values
- Remove unnecessary static_cast<int64_t> in serialization

Breaking change: API consumers expecting float will receive int

+++ include/dmcp/doom/commands.h
@@ -53,7 +53,7 @@ typedef struct {
 } dmcp_cmd_give_item_t;
 
 typedef struct {
-  float health;
+  int32_t health;
 } dmcp_cmd_set_health_t;
 
 typedef struct {
@@ -74,9 +74,9 @@ typedef struct {
 } dmcp_cmd_timescale_t;
 
 typedef struct {
-  int32_t target_tid;  // Target entity ID
-  float   damage;
-  char    damage_type[32];  // e.g., "Normal", "Fire", "Ice"
+  int32_t target_tid;
+  int32_t damage;
+  char    damage_type[32];
 } dmcp_cmd_damage_t;
 
 typedef struct {
